+++ src/stamps/entries.js
@@ -4,7 +4,7 @@ import sessionStamp from './session';
 import { grab } from '../apiHelper';
 
 function getPathWithQs(path, params = {}) {
-  const { id, typeId, alias, preview, at, offset, size } = params;
+  const { id, typeId, alias, typeAlias, preview, at, offset, size } = params;
   const qsParams = {};
   // The id array must be turned into CSV
   if (id && id.length) { qsParams.id = `${id.join(',')}`; }
@@ -19,7 +19,7 @@ function getPathWithQs(path, params = {}) {
     qsParams.at = at.toISOString();
   }
   // Add curated and non-curated params
-  const queryString = qs.stringify(Object.assign({}, qsParams, { typeId, offset, size }));
+  const queryString = qs.stringify(Object.assign({}, qsParams, { typeId, typeAlias, offset, size }));
   return `${path}?${queryString}`;
 }
 
@@ -32,13 +32,14 @@ const stamp = stampit()
 .methods({
   /**
    * Get all the content entries, based on the given parameters.
-   * **DO NOT** use several of id, alias and typeId at the same time - behaviour would be ungaranteed.
+   * **DO NOT** use several of id, alias, typeId and typeAlias at the same time - behaviour would be ungaranteed.
    * @param {object} [params] a parameters object
    * @param {boolean} [params.preview] when true, get the preview version
    * @param {string|date} [params.at] when given, get the version at the given time
    * @param {array} [params.id] an array of entry ids (strings)
    * @param {array} [params.alias] an array of entry aliases (strings)
    * @param {string} [params.typeId] only return entries matching this type id
+   * @param {string} [params.typeAlias] only return entries matching this typeAlias
    * @param {number|string} [params.size] limit to that many results per page (limits as per AppGrid API, currently 1 to 50, default 20)
    * @param {number|string} [params.offset] offset the result by that many pages
    * @return {promise}  a promise of an array of entries (objects)