+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # TODO
 
-- MINOR: add getEntryByTypeAlias when available in prod
 - MINOR: add the new bulk method to add applogs when available in prod
 - MINOR: basic robot detection with a default deviceId for them all ?
 - MINOR: have extra configuration params given go through the middleware so they are used in the factory (that will allow setting the gid when using the middleware, for instance)
 
 # Unreleased
 
+- MINOR: getEntries can now take a typeAlias parameter
 - PATCH: Changed the AppGrid REST API link to its new location
 - PATCH: doc: installing this SDK should usually be done with --save, not --save-dev